feat: support embedding encrypted psiphon config
We use an optional build tag to hide this configuration. When you
choose this configuration, you need to provide the encrypted config
as well as the corresponding decryption key.

This is not the final design. This is an interim design to start
working and experimenting with this functionality.

Part of https://github.com/ooni/probe/issues/985
