+++ internal/engine/session_nopsiphon.go
@@ -1,0 +1,14 @@
+// +build !ooni_psiphon_config
+
+package engine
+
+import "context"
+
+// FetchPsiphonConfig fetches psiphon config from the API.
+func (s *Session) FetchPsiphonConfig(ctx context.Context) ([]byte, error) {
+	clnt, err := s.NewOrchestraClient(ctx)
+	if err != nil {
+		return nil, err
+	}
+	return clnt.FetchPsiphonConfig(ctx)
+}
